# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ISANORTE / ISADECO API
  version: 0.0.1-SNAPSHOT
  description: |
    Contrato estático derivado de los Controllers y DTO, actualizado el 2026-09-20.
    No implica integración de Swagger/OpenAPI en Spring Boot. Security/JWT no está implementado;
    `x-access-intent` expresa el acceso previsto y no una protección efectiva.
...
servers:
  - url: http://localhost:8080
    description: Desarrollo local con el puerto predeterminado de Spring Boot
...
paths:
...
  /api/publico/contacto:
    post:
      tags: [Público]
      operationId: crearSolicitudContactoPublica
      x-access-intent: PUBLICO
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SolicitudContactoCreateRequest'}}}}
      responses: {'201': {description: Recibida, content: {application/json: {schema: {$ref: '#/components/schemas/SolicitudContactoPublicResponse'}}}}, '400': {$ref: '#/components/responses/BadRequest'}}
  /api/publico/sitios/{clave}:
    get:
      tags: [Público]
      operationId: obtenerSitioPublico
      x-access-intent: PUBLICO
      parameters: [{name: clave, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Sitio canónico, content: {application/json: {schema: {$ref: '#/components/schemas/PublicSiteResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/publico/sitios/{clave}/home:
    get:
      tags: [Público]
      operationId: obtenerHomePublica
      x-access-intent: PUBLICO
      parameters: [{name: clave, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Home agregada, content: {application/json: {schema: {$ref: '#/components/schemas/PublicHomeResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/publico/sitios/{clave}/paginas/{pagina}:
    get:
      tags: [Público]
      operationId: obtenerPaginaPublica
      x-access-intent: PUBLICO
      parameters:
        - {name: clave, in: path, required: true, schema: {type: string}}
        - {name: pagina, in: path, required: true, schema: {$ref: '#/components/schemas/TipoPaginaPublica'}}
      responses: {'200': {description: Editorial y SEO, content: {application/json: {schema: {$ref: '#/components/schemas/PublicPageResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/publico/sitios/{clave}/unidades/{slug}:
    get:
      tags: [Público]
      operationId: obtenerUnidadPublica
      x-access-intent: PUBLICO
      parameters:
        - {name: clave, in: path, required: true, schema: {type: string}}
        - {name: slug, in: path, required: true, schema: {type: string}}
      responses: {'200': {description: Unidad activa, content: {application/json: {schema: {$ref: '#/components/schemas/PublicBusinessUnitResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/publico/sitios/{clave}/unidades/{unidadSlug}/catalogo:
    get:
      tags: [Público]
      operationId: obtenerCatalogoPublicoPorUnidad
      x-access-intent: PUBLICO
      description: Catálogo web del dominio Producto, limitado a una unidad activa del sitio. No representa un recurso PDF de unidad.
      parameters:
        - {name: clave, in: path, required: true, schema: {type: string}}
        - {name: unidadSlug, in: path, required: true, schema: {type: string}}
      responses: {'200': {description: Catálogo de productos públicos de la unidad, posiblemente vacío, content: {application/json: {schema: {$ref: '#/components/schemas/PublicProductCatalogResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/publico/sitios/{clave}/unidades/{unidadSlug}/productos/{productoSlug}:
    get:
      tags: [Público]
      operationId: obtenerProductoPublicoPorUnidad
      x-access-intent: PUBLICO
      description: Detalle de un producto publicado cuya unidad activa pertenece al sitio solicitado.
      parameters:
        - {name: clave, in: path, required: true, schema: {type: string}}
        - {name: unidadSlug, in: path, required: true, schema: {type: string}}
        - {name: productoSlug, in: path, required: true, schema: {type: string}}
      responses: {'200': {description: Producto público de la unidad, content: {application/json: {schema: {$ref: '#/components/schemas/PublicProductDetailResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/publico/sitios/{clave}/unidades/{unidadSlug}/productos/{productoSlug}/comparar-precio:
    post:
      tags: [Público]
      operationId: compararPrecioProductoPublico
      x-access-intent: PUBLICO
      summary: Comparar precioBase con el precio extraído de una URL de producto
      description: |
        Consulta explícita sin persistencia ni historial. Valida sitio, empresa, unidad activa
        y producto publicado. La moneda interna de esta versión es PEN.
        El servidor obtiene precioBase; nunca acepta un precio interno del cliente.
        Solo HTTP/HTTPS, puertos estándar y direcciones públicas. Hasta 3 redirecciones
        validadas, sin downgrade HTTPS, HTML de hasta 1 MiB y plazo por salto de 8 segundos.
        Los fallos externos esperables se devuelven con HTTP 200 y un estado no SUCCESS.
        La identidad/equivalencia del producto externo debe comprobarla el usuario.
      parameters:
        - {name: clave, in: path, required: true, schema: {type: string}, example: isanorte}
        - {name: unidadSlug, in: path, required: true, schema: {type: string}, example: isadecor}
        - {name: productoSlug, in: path, required: true, schema: {type: string}, example: mesa}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ComparacionPrecioRequest'}
      responses:
        '200':
          description: Resultado de comparación o fallo externo controlado. Cache-Control no-store.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ComparacionPrecioResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/publico/sitios/{clave}/unidades/{unidadSlug}/productos/{productoSlug}/comparacion-competidores:
    get:
      tags: [Público]
      operationId: compararProductosCompetidores
      x-access-intent: PUBLICO
      summary: Busca productos similares en PISOPAK y DECORPLAS
      description: Consulta bajo acción explícita, sin persistencia. El backend recupera el producto publicado de ISADECOR, consulta únicamente dominios oficiales, puntúa candidatos y compara precios solo con moneda y unidad compatibles.
      parameters:
        - {name: clave, in: path, required: true, schema: {type: string}, example: isanorte}
        - {name: unidadSlug, in: path, required: true, schema: {type: string}, example: isadecor}
        - {name: productoSlug, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Resultado independiente por tienda. Cache-Control no-store.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ComparacionCompetidoresResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
components:
...
  responses:
    BadRequest:
      description: Request inválido, validación fallida o JSON no legible
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
    NotFound:
      description: Recurso o relación no encontrada
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
...
  schemas:
    EstadoComparacionCompetidor:
      type: string
      enum: [FOUND, NO_SIMILAR_PRODUCT_FOUND, STORE_UNAVAILABLE]
    CaracteristicaComparable:
      type: object
      required: [nombre, valor]
      properties:
        nombre: {type: string}
        valor: {type: string}
    ProductoComparable:
      type: object
      required: [nombre, precio, precioAnterior, moneda, unidadPrecio, caracteristicas]
      properties:
        nombre: {type: string}
        precio: {type: number, nullable: true}
        precioAnterior: {type: number, nullable: true}
        moneda: {type: string, example: PEN}
        unidadPrecio: {type: string, nullable: true, example: m2}
        caracteristicas: {type: array, items: {$ref: '#/components/schemas/CaracteristicaComparable'}}
    ResultadoCompetidor:
      type: object
      required: [empresa, estado, encontrado, nombreProducto, urlProducto, precio, precioAnterior, moneda, unidadPrecio, cantidadPorPresentacion, similitud, caracteristicas, comparablePrecio, diferenciaPrecio, mensaje]
      properties:
        empresa: {type: string, enum: [PISOPAK, DECORPLAS]}
        estado: {$ref: '#/components/schemas/EstadoComparacionCompetidor'}
        encontrado: {type: boolean}
        nombreProducto: {type: string, nullable: true}
        urlProducto: {type: string, format: uri, nullable: true}
        precio: {type: number, nullable: true}
        precioAnterior: {type: number, nullable: true}
        moneda: {type: string, nullable: true}
        unidadPrecio: {type: string, nullable: true}
        cantidadPorPresentacion: {type: number, nullable: true}
        similitud: {type: number, minimum: 0, maximum: 1, nullable: true}
        caracteristicas: {type: array, items: {$ref: '#/components/schemas/CaracteristicaComparable'}}
        comparablePrecio: {type: boolean}
        diferenciaPrecio: {type: number, nullable: true, description: 'precio competidor - precio ISADECOR; solo con moneda y unidad compatibles'}
        mensaje: {type: string}
    ComparacionCompetidoresResponse:
      type: object
      required: [productoIsadecor, competidores, diferenciasEncontradas, fechaConsulta]
      properties:
        productoIsadecor: {$ref: '#/components/schemas/ProductoComparable'}
        competidores: {type: array, items: {$ref: '#/components/schemas/ResultadoCompetidor'}}
        diferenciasEncontradas: {type: array, items: {type: string}}
        fechaConsulta: {type: string, format: date-time}
    ComparacionPrecioRequest:
      type: object
      description: La ausencia, vacío o formato inválido de URL devuelve INVALID_URL; más de 2048 caracteres devuelve HTTP 400.
      properties:
        urlExterna: {type: string, nullable: true, maxLength: 2048, example: 'https://tienda.example/productos/mesa'}
    EstadoComparacionPrecio:
      type: string
      enum: [SUCCESS, INVALID_URL, BLOCKED_URL, SITE_UNREACHABLE, UNSUPPORTED_CONTENT, PRICE_NOT_FOUND, CURRENCY_UNKNOWN, CURRENCY_MISMATCH, NOT_COMPARABLE]
    ComparacionPrecioResponse:
      type: object
      required: [producto, urlExterna, dominioExterno, nombreProductoExterno, precioInterno, precioExterno, monedaInterna, monedaExterna, diferencia, porcentajeDiferencia, comparable, estado, mensaje, fechaConsulta]
      properties:
        producto: {type: string}
        urlExterna: {type: string, nullable: true, description: URL final validada; null si falla la extracción}
        dominioExterno: {type: string, nullable: true}
        nombreProductoExterno: {type: string, nullable: true}
        precioInterno: {type: number, nullable: true, description: precioBase obtenido por el backend}
        precioExterno: {type: number, nullable: true}
        monedaInterna: {type: string, example: PEN}
        monedaExterna: {type: string, nullable: true}
        diferencia: {type: number, nullable: true, description: 'precioExterno - precioInterno; solo cuando comparable=true'}
        porcentajeDiferencia: {type: number, nullable: true, description: '(precioExterno - precioInterno) * 100 / precioInterno; null si no comparable o precioInterno=0. Redondeo HALF_UP a 2 decimales.'}
        comparable: {type: boolean}
        estado: {$ref: '#/components/schemas/EstadoComparacionPrecio'}
        mensaje: {type: string}
        fechaConsulta: {type: string, format: date-time}
...
    EstadoDisponibilidad:
      type: string
      enum: [DISPONIBLE, AGOTADO, BAJO_PEDIDO, CONSULTAR]
...
    TipoDocumentoProducto:
      type: string
      enum: [CATALOGO, FICHA_TECNICA, MANUAL, OTRO]
...
    TipoSeccionLanding:
      type: string
      enum: [HERO, EMPRESA, SERVICIOS, UNIDAD_NEGOCIO, PROYECTOS, CONTACTO, CTA, PERSONALIZADA]
...
    ApiErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp: {type: string, format: date-time, description: OffsetDateTime con offset}
        status: {type: integer, format: int32}
        error: {type: string}
        message: {type: string, nullable: true}
        path: {type: string}
      example:
        timestamp: '2026-09-13T18:30:00-05:00'
        status: 404
        error: Not Found
        message: 'Registro no encontrado con ID: 11111111-1111-4111-8111-111111111111'
        path: /api/productos/11111111-1111-4111-8111-111111111111
...
    TipoPaginaPublica:
      type: string
      enum: [NOSOTROS, SERVICIOS, PROYECTOS, CONTACTO]
...
    TipoRecursoUnidadNegocio:
      type: string
      enum: [IMAGEN_FONDO, IMAGEN_EDITORIAL, CATALOGO]
    RobotsSeo:
      type: string
      enum: [INDEX_FOLLOW, NOINDEX_FOLLOW, INDEX_NOFOLLOW, NOINDEX_NOFOLLOW]
    EstadoSolicitudContacto:
      type: string
      enum: [NUEVA, EN_GESTION, RESPONDIDA, DESCARTADA, SPAM]
...
    SolicitudContactoCreateRequest:
      type: object
      required: [nombre, email, telefono, mensaje]
      properties:
        nombre: {type: string, maxLength: 150}
        email: {type: string, format: email, maxLength: 180}
        telefono: {type: string, maxLength: 30}
        empresa: {type: string, maxLength: 180, nullable: true}
        mensaje: {type: string, maxLength: 5000}
...
    SolicitudContactoPublicResponse:
      type: object
      required: [id, estado, fechaCreacion]
      properties:
        id: {type: string, format: uuid}
        estado: {$ref: '#/components/schemas/EstadoSolicitudContacto'}
        fechaCreacion: {type: string, format: date-time}
...
    PublicSiteResponse:
      type: object
      description: Configuración, empresa sin RUC, redes activas y unidades activas del sitio canónico.
      required: [clave, empresa, redes, unidades]
      properties:
        clave: {type: string}
        tituloSitio: {type: string, nullable: true}
        descripcionSitio: {type: string, nullable: true}
        logoUrl: {type: string, nullable: true}
        logoBlancoUrl: {type: string, nullable: true}
        faviconUrl: {type: string, nullable: true}
        textoPiePagina: {type: string, nullable: true}
        empresa: {$ref: '#/components/schemas/PublicEmpresa'}
        redes: {type: array, items: {$ref: '#/components/schemas/PublicRedSocial'}}
        unidades: {type: array, items: {$ref: '#/components/schemas/PublicUnidadResumen'}}
    PublicEmpresa:
      type: object
      properties:
        nombreComercial: {type: string}
        direccion: {type: string, nullable: true}
        ciudad: {type: string, nullable: true}
        telefono: {type: string, nullable: true}
        telefonoSecundario: {type: string, nullable: true}
        email: {type: string, format: email, nullable: true}
        emailVentas: {type: string, format: email, nullable: true}
        whatsapp: {type: string, nullable: true}
        horarioAtencion: {type: string, nullable: true}
        resumenNosotros: {type: string, nullable: true}
    PublicRedSocial:
      type: object
      properties:
        nombre: {type: string}
        url: {type: string}
        icono: {type: string, nullable: true}
        orden: {type: integer}
    PublicUnidadResumen:
      type: object
      properties:
        nombre: {type: string}
        slug: {type: string}
        descripcion: {type: string, nullable: true}
        icono: {type: string, nullable: true}
        imagenUrl: {type: string, nullable: true}
        imagenAlt: {type: string, nullable: true}
        orden: {type: integer}
    PublicHomeResponse:
      type: object
      description: Home agregada; todas las colecciones están filtradas y ordenadas.
      required: [secciones, servicios, proyectos, unidadDestacada]
      properties:
        secciones: {type: array, items: {$ref: '#/components/schemas/PublicHomeSeccion'}}
        servicios: {type: array, items: {$ref: '#/components/schemas/PublicHomeServicio'}}
        proyectos: {type: array, items: {$ref: '#/components/schemas/PublicHomeProyecto'}}
        unidadDestacada:
          allOf: [{$ref: '#/components/schemas/PublicHomeUnidad'}]
          nullable: true
    PublicHomeSeccion:
      type: object
      properties:
        tipo: {$ref: '#/components/schemas/TipoSeccionLanding'}
        etiqueta: {type: string, nullable: true}
        titulo: {type: string, nullable: true}
        subtitulo: {type: string, nullable: true}
        contenido: {type: string, nullable: true}
        imagenUrl: {type: string, nullable: true}
        imagenAlt: {type: string, nullable: true}
        textoBoton: {type: string, nullable: true}
        enlaceBoton: {type: string, nullable: true}
        orden: {type: integer}
        escenas: {type: array, items: {$ref: '#/components/schemas/PublicHomeEscena'}}
        acciones: {type: array, items: {$ref: '#/components/schemas/PublicHomeAccion'}}
    PublicHomeEscena:
      type: object
      properties:
        imagenUrl: {type: string}
        alt: {type: string, nullable: true}
        orden: {type: integer}
    PublicHomeAccion:
      type: object
      properties:
        texto: {type: string}
        enlace: {type: string}
        orden: {type: integer}
    PublicHomeServicio:
      type: object
      properties:
        nombre: {type: string}
        slug: {type: string}
        resumen: {type: string, nullable: true}
        descripcion: {type: string}
        icono: {type: string, nullable: true}
        imagenUrl: {type: string, nullable: true}
        imagenAlt: {type: string, nullable: true}
        etiqueta: {type: string, nullable: true}
        orden: {type: integer}
        beneficios: {type: array, items: {type: string}}
    PublicHomeProyecto:
      type: object
      properties:
        nombre: {type: string}
        slug: {type: string}
        ubicacion: {type: string, nullable: true}
        fechaProyecto: {type: string, nullable: true}
        descripcion: {type: string}
        orden: {type: integer}
        imagenes: {type: array, items: {$ref: '#/components/schemas/PublicHomeImagenProyecto'}}
    PublicHomeImagenProyecto:
      type: object
      properties:
        url: {type: string}
        alt: {type: string, nullable: true}
        esPrincipal: {type: boolean}
        orden: {type: integer}
    PublicRecursoUnidad:
      type: object
      properties:
        tipo: {$ref: '#/components/schemas/TipoRecursoUnidadNegocio'}
        url: {type: string}
        alt: {type: string, nullable: true}
        etiqueta: {type: string, nullable: true}
        orden: {type: integer}
    PublicHomeUnidad:
      allOf:
        - {$ref: '#/components/schemas/PublicUnidadResumen'}
        - type: object
          properties:
            recursos: {type: array, items: {$ref: '#/components/schemas/PublicRecursoUnidad'}}
    PublicPageResponse:
      type: object
      required: [contenido, seo]
      properties:
        contenido: {$ref: '#/components/schemas/PublicContenidoPagina'}
        seo:
          allOf: [{$ref: '#/components/schemas/PublicSeo'}]
          nullable: true
        empresa:
          allOf: [{$ref: '#/components/schemas/PublicEmpresaNosotros'}]
          nullable: true
        servicios:
          type: array
          nullable: true
          items: {$ref: '#/components/schemas/PublicServiceResponse'}
        proyectos:
          type: array
          nullable: true
          items: {$ref: '#/components/schemas/PublicProjectResponse'}
    PublicContenidoPagina:
      type: object
      properties:
        pagina: {$ref: '#/components/schemas/TipoPaginaPublica'}
        eyebrow: {type: string, nullable: true}
        titulo: {type: string, nullable: true}
        introduccion: {type: string, nullable: true}
        descripcion: {type: string, nullable: true}
        imagenUrl: {type: string, nullable: true}
        imagenAlt: {type: string, nullable: true}
        imagenFondoUrl: {type: string, nullable: true}
        tags: {type: array, items: {type: string}}
    PublicSeo:
      type: object
      properties:
        title: {type: string}
        description: {type: string}
        ogImageUrl: {type: string, nullable: true}
        robots: {$ref: '#/components/schemas/RobotsSeo'}
    PublicEmpresaNosotros:
      description: Datos corporativos expuestos únicamente al solicitar la página NOSOTROS.
      properties:
        mision: {type: string, nullable: true}
        vision: {type: string, nullable: true}
        valores: {type: string, nullable: true}
        estadisticas:
          type: array
          items: {$ref: '#/components/schemas/PublicEstadisticaEmpresa'}
    PublicEstadisticaEmpresa:
      properties:
        valor: {type: number, format: decimal}
        prefijo: {type: string, nullable: true}
        sufijo: {type: string, nullable: true}
        etiqueta: {type: string}
        orden: {type: integer}
    PublicServiceResponse:
      description: Servicio editorial público usado por la página SERVICIOS.
      properties:
        nombre: {type: string}
        slug: {type: string}
        etiqueta: {type: string, nullable: true}
        resumen: {type: string, nullable: true}
        descripcion: {type: string, nullable: true}
        imagenUrl: {type: string, nullable: true}
        imagenAlt: {type: string, nullable: true}
        orden: {type: integer}
        beneficios:
          type: array
          items: {$ref: '#/components/schemas/PublicServiceBenefitResponse'}
    PublicServiceBenefitResponse:
      properties:
        texto: {type: string}
        orden: {type: integer}
    PublicProjectResponse:
      description: Proyecto editorial público usado por la página PROYECTOS.
      properties:
        nombre: {type: string}
        slug: {type: string}
        descripcion: {type: string, nullable: true}
        ubicacion: {type: string, nullable: true}
        fechaProyecto: {type: string, nullable: true, description: Texto editorial libre del dominio actual.}
        orden: {type: integer}
        imagenes:
          type: array
          items: {$ref: '#/components/schemas/PublicProjectImageResponse'}
        servicios:
          type: array
          items: {$ref: '#/components/schemas/PublicProjectServiceResponse'}
    PublicProjectImageResponse:
      properties:
        url: {type: string}
        alt: {type: string, nullable: true}
        esPrincipal: {type: boolean}
        orden: {type: integer}
    PublicProjectServiceResponse:
      properties:
        nombre: {type: string}
        slug: {type: string}
    PublicBusinessUnitResponse:
      type: object
      description: Unidad activa perteneciente a la empresa del sitio, recursos activos y SEO opcional.
      properties:
        nombre: {type: string}
        slug: {type: string}
        descripcion: {type: string, nullable: true}
        icono: {type: string, nullable: true}
        imagenUrl: {type: string, nullable: true}
        imagenAlt: {type: string, nullable: true}
        recursos: {type: array, items: {$ref: '#/components/schemas/PublicRecursoUnidad'}}
        seo:
          allOf: [{$ref: '#/components/schemas/PublicSeo'}]
          nullable: true
    PublicProductCatalogResponse:
      type: object
      description: Catálogo web del dominio Producto de una unidad activa; no es el recurso PDF CATALOGO.
      required: [unidad, categorias, productos]
      properties:
        unidad: {$ref: '#/components/schemas/PublicProductCatalogUnitResponse'}
        categorias:
          type: array
          items: {$ref: '#/components/schemas/PublicProductCategoryResponse'}
        productos:
          type: array
          items: {$ref: '#/components/schemas/PublicProductCardResponse'}
    PublicProductCatalogUnitResponse:
      type: object
      required: [nombre, slug]
      properties:
        nombre: {type: string}
        slug: {type: string}
    PublicProductCardResponse:
      type: object
      required: [nombre, sku, slug, descripcion, disponibilidad, categorias]
      properties:
        nombre: {type: string}
        sku: {type: string}
        slug: {type: string}
        resumen: {type: string, nullable: true}
        descripcion: {type: string}
        precioBase: {type: number, format: decimal, nullable: true}
        precioAnterior: {type: number, format: decimal, nullable: true}
        descuentoPorcentaje: {type: number, format: decimal, nullable: true}
        disponibilidad: {$ref: '#/components/schemas/EstadoDisponibilidad'}
        retiroEnTienda: {type: boolean, nullable: true}
        imagen:
          allOf: [{$ref: '#/components/schemas/PublicProductImageResponse'}]
          nullable: true
        categorias:
          type: array
          items: {$ref: '#/components/schemas/PublicProductCategoryResponse'}
    PublicProductDetailResponse:
      type: object
      required: [nombre, sku, slug, descripcion, disponibilidad, categorias, imagenes, variantes, especificaciones, documentos]
      properties:
        nombre: {type: string}
        sku: {type: string}
        slug: {type: string}
        resumen: {type: string, nullable: true}
        descripcion: {type: string}
        tituloSeo: {type: string, nullable: true}
        descripcionSeo: {type: string, nullable: true}
        precioBase: {type: number, format: decimal, nullable: true}
        precioAnterior: {type: number, format: decimal, nullable: true}
        descuentoPorcentaje: {type: number, format: decimal, nullable: true}
        disponibilidad: {$ref: '#/components/schemas/EstadoDisponibilidad'}
        retiroEnTienda: {type: boolean, nullable: true}
        categorias:
          type: array
          items: {$ref: '#/components/schemas/PublicProductCategoryResponse'}
        imagenes:
          type: array
          items: {$ref: '#/components/schemas/PublicProductImageResponse'}
        variantes:
          type: array
          items: {$ref: '#/components/schemas/PublicProductVariantResponse'}
        especificaciones:
          type: array
          items: {$ref: '#/components/schemas/PublicProductSpecificationResponse'}
        documentos:
          type: array
          items: {$ref: '#/components/schemas/PublicProductDocumentResponse'}
        configuracionCalculo:
          allOf: [{$ref: '#/components/schemas/PublicProductCalculationResponse'}]
          nullable: true
    PublicProductCategoryResponse:
      type: object
      required: [nombre, slug]
      properties:
        nombre: {type: string}
        slug: {type: string}
    PublicProductImageResponse:
      type: object
      required: [url, esPrincipal, orden]
      properties:
        url: {type: string}
        alt: {type: string, nullable: true}
        esPrincipal: {type: boolean}
        orden: {type: integer, format: int32}
    PublicProductVariantResponse:
      type: object
      required: [sku, nombre, disponible, orden]
      properties:
        sku: {type: string}
        nombre: {type: string}
        descripcion: {type: string, nullable: true}
        precio: {type: number, format: decimal, nullable: true}
        disponible: {type: boolean}
        imagenUrl: {type: string, nullable: true}
        orden: {type: integer, format: int32}
    PublicProductSpecificationResponse:
      type: object
      required: [clave, valor, orden]
      properties:
        clave: {type: string}
        valor: {type: string}
        grupo: {type: string, nullable: true}
        orden: {type: integer, format: int32}
    PublicProductDocumentResponse:
      type: object
      required: [titulo, url, tipo]
      properties:
        titulo: {type: string}
        url: {type: string}
        tipo: {$ref: '#/components/schemas/TipoDocumentoProducto'}
        formato: {type: string, nullable: true}
        tamanoBytes: {type: integer, format: int64, nullable: true}
    PublicProductCalculationResponse:
      type: object
      required: [habilitada, coberturaPorUnidad]
      properties:
        habilitada: {type: boolean}
        etiquetaEntrada: {type: string, nullable: true}
        unidadEntrada: {type: string, nullable: true}
        coberturaPorUnidad: {type: number, format: decimal}
        unidadVenta: {type: string, nullable: true}
        textoAyuda: {type: string, nullable: true}
...